requirements.txt
fastapi
uvicorn
chromadb
sentence-transformers
python-multipart
Pillow
